Rename modules in server package to make more sense
