+++ config_db_migrate/env.py
@@ -9,7 +9,7 @@ config = context.config
 # Add your model's MetaData object here
 # for 'autogenerate' support.
 try:
-    from libcodechecker.server.config_db_model import Base
+    from libcodechecker.server.database.config_db_model import Base
 except ImportError:
     # Assume we are in the source directory
     import sys
