+++ libcodechecker/server/api/authentication.py
